utterly broken, working on v2 concepts
