+++ go.mod
@@ -1,19 +1,11 @@
-module github.com/myENA/consultant
+module github.com/myENA/consultant/v2
+
+go 1.12
 
 require (
-	github.com/armon/go-metrics v0.0.0-20190423201044-2801d9688273 // indirect
-	github.com/hashicorp/consul/api v1.1.0
-	github.com/hashicorp/consul/sdk v0.1.1
-	github.com/hashicorp/go-msgpack v0.5.4 // indirect
-	github.com/hashicorp/go-sockaddr v1.0.2 // indirect
-	github.com/hashicorp/golang-lru v0.5.1 // indirect
-	github.com/miekg/dns v1.1.8 // indirect
-	github.com/mitchellh/go-homedir v1.1.0 // indirect
+	github.com/hashicorp/consul/api v1.3.0
+	github.com/hashicorp/consul/sdk v0.3.0
 	github.com/myENA/go-helpers v0.1.0
 	github.com/pkg/errors v0.8.1
-	github.com/stretchr/testify v1.3.0
-	golang.org/x/crypto v0.0.0-20190422183909-d864b10871cd // indirect
-	golang.org/x/net v0.0.0-20190424024845-afe8014c977f // indirect
-	golang.org/x/sync v0.0.0-20190423024810-112230192c58 // indirect
-	golang.org/x/sys v0.0.0-20190422165155-953cdadca894 // indirect
+	github.com/stretchr/testify v1.4.0
 )
