--- log/log_test.go
@@ -1,7 +1,0 @@
-package log_test
-
-import "testing"
-
-func TestLog(t *testing.T) {
-	// TODO
-}
