+++ sibling_locator_test.go
@@ -3,8 +3,8 @@ package consultant_test
 import (
 	"fmt"
 	"github.com/hashicorp/consul/api"
-	"github.com/myENA/consultant"
-	"github.com/myENA/consultant/testutil"
+
+	"github.com/myENA/consultant/v2"
 	"github.com/stretchr/testify/require"
 	"github.com/stretchr/testify/suite"
 	"log"
