--- testutil/testutil_test.go
@@ -1,7 +1,0 @@
-package testutil_test
-
-import "testing"
-
-func TestTestUtil(t *testing.T) {
-	// TODO
-}
